application summary
Dear all,
my dream ever was being a programmer. I spent and exerted so much effort in this app.
I know it not completed but I uploaded it so you know am trying my best to finish the project.
my the obligations of life, work, family & this scholarship is so much overwhelming.
I made the main table enabling the entry of items, item name, amount & count - and summing the total in the same table
and totaling the whole invoice in a label above.
I hope you would review the code and give me hints to continue the rest of missing parts.

thanks a lot